Trait forest_chain::Scale

pub trait Scale {
    fn weight<DB>(db: &DB, ts: &Tipset) -> Result<Weight, Error>
    where
        DB: BlockStore;
}

The Scale trait abstracts away the logic of assigning a weight to a chain, which can be consensus specific. For example it can depend on the stake and power of validators, or it can be as simple as the height of the blocks in a Nakamoto style consensus.

Required Methods

fn weight<DB>(db: &DB, ts: &Tipset) -> Result<Weight, Error>where
DB: BlockStore,

Calculate the weight of a tipset.